lift ElectronicEigenvalues, change n_bands to n_levels
